:bug: Fix incrorrectly formed column

+++ database/migrations/2022_09_26_123543_add_website_article_reactions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('website_article_reactions', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('article_id');
+            $table->unsignedBigInteger('article_id');
             $table->string('reaction', 50);
             $table->boolean('active')->default(true);
             $table->foreign('article_id')->references('id')->on('website_articles')->cascadeOnDelete();
